requirements/dev.txt
epc
jedi
pudb
pytest==2.9.2
pytest-django==2.9.1
pytest-xdist==1.14
